Backend routes protected

+++ backend/src/controllers/notesControllers.ts
@@ -2,10 +2,13 @@ import { RequestHandler } from "express";
 import NoteModel from '../models/note'
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId
     try{
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId)
+        const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(notes)
     }
     catch(error){
@@ -15,13 +18,18 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getSingleNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId
+    const authenticatedUserId = req.session.userId
     try {
+        assertIsDefined(authenticatedUserId)
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid NoteId")
         }
         const note = await NoteModel.findById(noteId).exec();
         if(!note){
             throw createHttpError(404, "Note not found")
+        }
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You can not access this note.")
         }
         res.status(200).json(note)
     } catch (error) {
@@ -37,11 +45,14 @@ interface CreateNoteBody{
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title
     const text = req.body.text
+    const authenticatedUserId = req.session.userId
     try{
+        assertIsDefined(authenticatedUserId)
         if(!title){
             throw createHttpError(400, "Note must have a title")
         }
         const newNote =  await NoteModel.create({
+            userId: authenticatedUserId,
             title: title, 
             text: text
         })
@@ -65,8 +76,12 @@ export const updateNote: RequestHandler<UpdateNoteParam, unknown, UpdateNoteBody
     const noteId = req.params.noteId
     const newTitle = req.body.title
     const newText = req.body.text
+    const authenticatedUserId = req.session.userId
+
 
     try {
+        assertIsDefined(authenticatedUserId)
+
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid NoteId")
         }
@@ -78,6 +93,10 @@ export const updateNote: RequestHandler<UpdateNoteParam, unknown, UpdateNoteBody
         
         if(!note){
             throw createHttpError(404, "Note not found")
+        }
+
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You can not access this note.")
         }
 
         note.title = newTitle
@@ -92,14 +111,20 @@ export const updateNote: RequestHandler<UpdateNoteParam, unknown, UpdateNoteBody
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId
+    const authenticatedUserId = req.session.userId
 
     try {
+        assertIsDefined(authenticatedUserId)
+
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid NoteId")
         }
         const note = await NoteModel.findById(noteId).exec()
         if(!note){
             throw createHttpError(404, "Note not found")
+        }
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You can not access this note.")
         }
         await NoteModel.findByIdAndDelete(noteId)
         res.sendStatus(204)
